Create AuthValidator service, move AuthGuard validation logic to new service
