+++ src/services/AuthValidator/IAuthValidator.ts
@@ -1,0 +1,5 @@
+import { Request } from "express";
+
+export default interface IAuthValidator {
+    validateAndDecodeToken(request: Request): any;
+}
