+++ src/services/AuthValidator/index.ts
@@ -1,0 +1,1 @@
+export { AuthValidator } from "./AuthValidator";
